+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- ids parameter to transaction.query
+- ids parameter to transfer.query
 
 ## [0.6.0] - 2020-07-12
 ### Added